requirements.txt
Django
Pillow
djangorestframework
